+++ src/Components/Home.js
@@ -1,13 +1,13 @@
 import '../Components/home.css'
 
 
-import Bouncerate from './Charts/BounceRate';
-import FormSubmit from './Charts/FormSubmit';
-import RegionData from './Charts/RegionData';
-import PagePath from './Charts/PagePath';
-import TotalUsersCount from './Charts/TotalUsersCount';
-import DonatedUsersData from './Charts/DonatedUsersChart';
-import MonthlyReport from './Charts/MonthlyReport';
+import Bouncerate from './Charts/Report_Data/BounceRate';
+import FormSubmit from './Charts/Report_Data/FormSubmit';
+import RegionData from './Charts/Report_Data/RegionData';
+import PagePath from './Charts/Report_Data/PagePath';
+import TotalUsersCount from './Charts/Report_Data/TotalUsersCount';
+import DonatedUsersData from './Charts/Report_Data/DonatedUsersChart';
+import MonthlyReport from './Charts/Report_Data/MonthlyReport';
 
 
 const Home = () => {
