+++ src/Components/ReamTimeHome.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+function ReamTimeHome() {
+  return (
+    <div>
+    <div>
+    <h1 style={{ textAlign: 'center' }}> Real Time Data Charts</h1>
+    </div>
+      
+    </div>
+  )
+}
+
+export default ReamTimeHome;
